# apps/api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shelf API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
      "def-0": {
        "type": "object",
        "required": [
          "apiVersion",
          "kind",
          "workspaceId",
          "artifactId",
          "revisionId",
          "contentHash",
          "byteCount",
          "fileCount",
          "provenance",
          "publisherMetadata",
          "requestId",
          "paths",
          "replayed"
        ],
        "properties": {
          "apiVersion": {
            "type": "string",
            "enum": ["v1"]
          },
          "kind": {
            "type": "string",
            "enum": ["file"]
          },
          "workspaceId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "artifactId": {
            "type": "string",
            "pattern": "^art_[A-Za-z0-9_-]{22}$",
            "description": "Opaque artifact identifier containing 128 bits of entropy."
          },
          "revisionId": {
            "type": "string",
            "pattern": "^rev_[A-Za-z0-9_-]{22}$",
            "description": "Opaque revision identifier containing 128 bits of entropy."
          },
          "contentHash": {
            "type": "string",
            "pattern": "^sha256:[a-f0-9]{64}$"
          },
          "byteCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "fileCount": {
            "type": "number",
            "enum": [1]
          },
          "provenance": {
            "type": "object",
            "required": ["classification", "observed"],
            "properties": {
              "classification": {
                "type": "string",
                "enum": ["direct-publish"]
              },
              "observed": {
                "type": "object",
                "required": ["actorId", "operation"],
                "properties": {
                  "actorId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "operation": {
                    "type": "string",
                    "enum": ["file.publish"]
                  }
                },
                "additionalProperties": false
              }
            },
            "additionalProperties": false
          },
          "publisherMetadata": {
            "type": "object",
            "maxProperties": 32,
            "description": "Untrusted publisher-supplied string metadata.",
            "additionalProperties": {
              "type": "string",
              "maxLength": 2048
            }
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "paths": {
            "type": "object",
            "required": ["artifact", "revision", "content"],
            "properties": {
              "artifact": {
                "type": "string",
                "pattern": "^/api/v1/artifacts/[^/]+$"
              },
              "revision": {
                "type": "string",
                "pattern": "^/api/v1/revisions/[^/]+$"
              },
              "content": {
                "type": "string",
                "pattern": "^/api/v1/revisions/[^/]+/content$"
              }
            },
            "additionalProperties": false
          },
          "replayed": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "title": "PublishResult"
      },
...
      "def-3": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "retryable", "requestId"],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "AUTHENTICATION_REQUIRED",
                  "AUTHORIZATION_DENIED",
                  "INVALID_REQUEST",
                  "IDEMPOTENCY_CONFLICT",
                  "ARTIFACT_NOT_FOUND",
                  "REVISION_NOT_FOUND",
                  "SHARE_NOT_FOUND",
                  "ACCESS_CREDENTIAL_NOT_FOUND",
                  "RANGE_NOT_SATISFIABLE",
                  "MULTI_RANGE_UNSUPPORTED",
                  "REQUEST_CANCELLED",
                  "CONTENT_UNAVAILABLE",
                  "SERVICE_UNAVAILABLE",
                  "INTERNAL_ERROR"
                ],
                "title": "ErrorCode"
              },
              "message": {
                "type": "string",
                "minLength": 1,
                "maxLength": 512
              },
              "retryable": {
                "type": "boolean"
              },
              "requestId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 128
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": ["reason"],
                  "properties": {
                    "field": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "reason": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 256
                    }
                  },
                  "additionalProperties": false,
                  "title": "ErrorDetail"
                },
                "maxItems": 32
              }
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false,
        "title": "ErrorEnvelope"
      },
...
    }
  },
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}/artifacts/{artifactId}/revisions": {
      "post": {
        "operationId": "publishArtifactRevisionV1",
        "summary": "Publish another immutable revision to one artifact",
        "tags": ["artifacts"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "publisherMetadata": {
                    "type": "string",
                    "description": "JSON object of string values. Must precede the file part."
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "isFile": true
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "workspaceId",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^art_[A-Za-z0-9_-]{22}$"
            },
            "in": "path",
            "name": "artifactId",
            "required": true,
            "description": "Opaque artifact identifier containing 128 bits of entropy."
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "499": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workspaces/{workspaceId}/folders/{artifactId}/revisions": {
      "post": {
        "operationId": "publishFolderRevisionV1",
        "summary": "Publish another complete snapshot to one folder artifact",
        "tags": ["folders"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["manifest"],
                "properties": {
                  "publisherMetadata": {
                    "type": "string",
                    "description": "JSON object of string values. Must precede file parts."
                  },
                  "manifest": {
                    "type": "string",
                    "description": "shelf-folder-manifest/v1 JSON. Must precede every ordered file part."
                  },
                  "file": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary",
                      "isFile": true
                    },
                    "maxItems": 1000
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "workspaceId",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^art_[A-Za-z0-9_-]{22}$"
            },
            "in": "path",
            "name": "artifactId",
            "required": true,
            "description": "Opaque artifact identifier containing 128 bits of entropy."
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "apiVersion",
                    "kind",
                    "workspaceId",
                    "artifactId",
                    "revisionId",
                    "contentHash",
                    "byteCount",
                    "fileCount",
                    "provenance",
                    "publisherMetadata",
                    "requestId",
                    "paths",
                    "replayed"
                  ],
                  "properties": {
                    "apiVersion": {
                      "type": "string",
                      "enum": ["v1"]
                    },
                    "kind": {
                      "type": "string",
                      "enum": ["folder"]
                    },
                    "workspaceId": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "artifactId": {
                      "type": "string",
                      "pattern": "^art_[A-Za-z0-9_-]{22}$",
                      "description": "Opaque artifact identifier containing 128 bits of entropy."
                    },
                    "revisionId": {
                      "type": "string",
                      "pattern": "^rev_[A-Za-z0-9_-]{22}$",
                      "description": "Opaque revision identifier containing 128 bits of entropy."
                    },
                    "contentHash": {
                      "type": "string",
                      "pattern": "^sha256:[a-f0-9]{64}$"
                    },
                    "byteCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "fileCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000
                    },
                    "provenance": {
                      "type": "object",
                      "required": ["classification", "observed"],
                      "properties": {
                        "classification": {
                          "type": "string",
                          "enum": ["direct-publish"]
                        },
                        "observed": {
                          "type": "object",
                          "required": ["actorId", "operation"],
                          "properties": {
                            "actorId": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 128
                            },
                            "operation": {
                              "type": "string",
                              "enum": ["file.publish"]
                            }
                          },
                          "additionalProperties": false
                        }
                      },
                      "additionalProperties": false
                    },
                    "publisherMetadata": {
                      "type": "object",
                      "maxProperties": 32,
                      "description": "Untrusted publisher-supplied string metadata.",
                      "additionalProperties": {
                        "type": "string",
                        "maxLength": 2048
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "paths": {
                      "type": "object",
                      "required": ["artifact", "revision", "tree"],
                      "properties": {
                        "artifact": {
                          "type": "string",
                          "pattern": "^/api/v1/artifacts/[^/]+$"
                        },
                        "revision": {
                          "type": "string",
                          "pattern": "^/api/v1/revisions/[^/]+$"
                        },
                        "tree": {
                          "type": "string",
                          "pattern": "^/api/v1/revisions/[^/]+/tree$"
                        }
                      },
                      "additionalProperties": false
                    },
                    "replayed": {
                      "type": "boolean"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "499": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/revisions/{baseRevisionId}/comparisons/{targetRevisionId}": {
      "get": {
        "operationId": "compareRevisionsV1",
        "summary": "Compare two immutable revisions of one artifact",
        "tags": ["revisions"],
        "description": "Compares immutable descriptors without opening content. Folder changes are paged; exact unambiguous byte matches identify moved files.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^rev_[A-Za-z0-9_-]{22}$"
            },
            "in": "path",
            "name": "baseRevisionId",
            "required": true,
            "description": "Opaque revision identifier containing 128 bits of entropy."
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^rev_[A-Za-z0-9_-]{22}$"
            },
            "in": "path",
            "name": "targetRevisionId",
            "required": true,
            "description": "Opaque revision identifier containing 128 bits of entropy."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-6"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          }
        }
      }
    },
...
  }
}